Decompose review runner into staged helpers
